adding tustin speakername to msgbox

+++ include/graphics.h
@@ -2,8 +2,8 @@
 #define GUARD_GRAPHICS_H
 
 // overworld
-extern const u8 gSignpostWindow_Gfx[];
-extern const u8 gMessageBox_Gfx[];
+extern const u32 gSignpostWindow_Gfx[];
+extern const u32 gMessageBox_Gfx[];
 extern const u16 gMessageBox_Pal[];
 
 // pokeballs
